feat: move configuration file to config folder, refactor what is launched when

+++ installers/fonts.sh
@@ -5,5 +5,14 @@
 #################
 
 # ComicShanns
-rm -rf $HOME/Library/Fonts/ComicShanns
-cp -r $DIR/fonts/ComicShanns $HOME/Library/Fonts/ComicShanns
+
+if [ ! $OS = "macos" ]; then
+    echo "Skipping fonts installation for non-macOS systems. Install the fonts manually."
+    echo "Press any key to continue"
+    read -k 1
+else
+    rm -rf $HOME/Library/Fonts/ComicShanns
+    cp -r $DIR/fonts/ComicShanns $HOME/Library/Fonts/ComicShanns
+    echo "Installed ComicShanns font"
+fi
+echo ""
